+++ TinkoffASDKCore/TinkoffASDKCore/Domain/AcquiringAPI/Requests/Implementations/AddCardRequest.swift
@@ -24,6 +24,7 @@ struct AddCardRequest: AcquiringRequest {
     let path: String = "v2/AddCard"
     let httpMethod: HTTPMethod = .post
     let parameters: HTTPParameters
+    let tokenFormationStrategy: TokenFormationStrategy = .includeAll()
 
     init(initAddCardData: AddCardData, baseURL: URL) {
         self.baseURL = baseURL