+++ TinkoffASDKCore/TinkoffASDKCore/Domain/AcquiringAPI/Requests/Implementations/GetTinkoffLinkRequest.swift
@@ -11,6 +11,7 @@ struct GetTinkoffLinkRequest: AcquiringRequest {
     let baseURL: URL
     let path: String
     let httpMethod: HTTPMethod = .get
+    let tokenFormationStrategy: TokenFormationStrategy = .none
 
     // MARK: - Init